Refactor org visas to restful routes
