+++ app/views/publish/providers/details.html.erb
@@ -74,7 +74,7 @@
              student_visa_sponsorship_status(@provider),
              %w[can_sponsor_student_visa],
              truncate_value: false,
-             action_path: @provider.provider_type == "university" ? student_visa_publish_provider_recruitment_cycle_path(@provider.provider_code, @provider.recruitment_cycle_year) : nil,
+             action_path: @provider.university? ? student_visa_publish_provider_recruitment_cycle_path(@provider.provider_code, @provider.recruitment_cycle_year) : nil,
              action_visually_hidden_text: "if candidates can get Student visa sponsorship",
            ) %>
