+++ config/routes/publish.rb
@@ -188,10 +188,10 @@ namespace :publish, as: :publish do
 
         get "/contact", on: :member, to: "contacts#edit"
         put "/contact", on: :member, to: "contacts#update"
-        get "/student-visa", on: :member, to: "visas#student_edit"
-        get "/skilled-worker-visa", on: :member, to: "visas#skilled_worker_edit"
-        put "/student-visa", on: :member, to: "visas#student_update"
-        put "/skilled-worker-visa", on: :member, to: "visas#skilled_worker_update"
+        get "/student-visa", on: :member, to: "student_visa#edit"
+        get "/skilled-worker-visa", on: :member, to: "skilled_worker_visa#edit"
+        put "/student-visa", on: :member, to: "student_visa#update"
+        put "/skilled-worker-visa", on: :member, to: "skilled_worker_visa#update"
 
         resources :allocations, only: %i[index], on: :member, param: :training_provider_code do
           collection do
